extract BaseTreeRewriter to be a clean base rewriter for expression rewrite
